agent-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1

langchain==0.3.14
langchain-anthropic==0.3.7
langchain-community==0.3.14

anthropic==0.45.0